fix: Ownership transfer should work right now. update: services to include security measures for inscription transfer.
Including new scripts for ownership transfer and security checks. The
ownership can be verified for selected TX-IDs. The security checks,
whether the inscription has enough confirmations in order to be
transfered. The fee rate was adapted, in order to be high enough and
future inscriptions might be made via the corresponding script
(test-inscription.mjs) by providing a file path. The ownership of
inscriptions is locked into 100 sats UTXOs, which should be big enough
to include the transfer for later (might be adapted in future).